Fix a BUNCH of things, feature flag redirects, change password handlers, use set_password instead
