+++ app/controllers/application_controller.rb
@@ -11,7 +11,7 @@ class ApplicationController < ActionController::Base
   fine_print_require :general_terms_of_use, :privacy_policy, unless: :disable_fine_print
 
   def newflow_feature_flag
-    if Settings::Db.store.newflow_feature_flag && !params[:bpff].present? # bpff = bypass feature flag
+    if request.get? && Settings::Db.store.newflow_feature_flag && !params[:bpff].present? # bpff = bypass feature flag
       redirect_to newflow_login_path(request.query_parameters)
     end
   end
