+++ spec/handlers/newflow/set_password_spec.rb
@@ -1,0 +1,8 @@
+require 'rails_helper'
+
+module Newflow
+  RSpec.describe SetPassword, type: :handler do
+    # TODO
+    it 'TODO'
+  end
+end
