--- spec/handlers/newflow/setup_password_spec.rb
@@ -1,7 +1,0 @@
-require 'rails_helper'
-
-module Newflow
-  RSpec.describe SetupPassword, type: :handler do
-    # TODO
-  end
-end
